requirements.txt
cryptography
django-cors-headers
drf-jwt
pyjwt
requests
